+++ packages/itwin/reports-config-widget/src/test/BulkExtractor.test.ts
@@ -2,11 +2,11 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import BulkExtractor from "../widget/components/BulkExtractor";
+import { BulkExtractor } from "../widget/components/BulkExtractor";
 import { ExtractionStates } from "../widget/components/ExtractionStatus";
 import { assert } from "chai";
 import * as moq from "typemoq";
-import { ExtractorState } from "@itwin/insights-client";
+import { ExtractionClient, ExtractorState, ReportsClient } from "@itwin/insights-client";
 import type { AccessToken } from "@itwin/core-bentley";
 
 jest.mock("../widget/components/Constants.ts", () => ({
@@ -40,6 +40,8 @@ afterEach(() => {
 
 const mockToastCallback = jest.fn();
 
+const mockGetAccessToken = async () => "mockAccessToken";
+
 const mockIModelId = "mockIModelId";
 const mockIModelId1 = "mockIModelId1";
 const mockReportId = "mockReportId";
@@ -47,13 +49,9 @@ const mockRunId = "mockRunId";
 const mockRunId1 = "mockRunId1";
 
 describe("BulkExtractor", () => {
-  const mockApiConfig = {
-    getAccessToken: async () => "mockAccessToken",
-    baseUrl: "",
-  };
 
   it("should return status none for unknown iModel", async () => {
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
     const result = await sut.getIModelState(mockIModelId, "", "");
     assert.strictEqual(result, ExtractionStates.None);
   });
@@ -67,7 +65,7 @@ describe("BulkExtractor", () => {
       async () => ({ state: ExtractorState.Running })
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
     await sut.runIModelExtraction(mockIModelId);
     const result = await sut.getIModelState(mockIModelId, "", "");
     assert.strictEqual(result, ExtractionStates.Running);
@@ -82,7 +80,7 @@ describe("BulkExtractor", () => {
       async () => ({ state: ExtractorState.Failed })
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
     await sut.runIModelExtraction(mockIModelId);
     const result = await sut.getIModelState(mockIModelId, "", "");
     assert.strictEqual(result, ExtractionStates.Failed);
@@ -105,7 +103,7 @@ describe("BulkExtractor", () => {
       async () => ({ state: ExtractorState.Succeeded })
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
     await sut.runIModelExtraction(mockIModelId);
 
     let result = await sut.getIModelState(mockIModelId, "", "");
@@ -137,7 +135,7 @@ describe("BulkExtractor", () => {
       async () => ({ state: ExtractorState.Queued })
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
 
     await sut.runIModelExtractions(mockIModelIds);
 
@@ -149,7 +147,7 @@ describe("BulkExtractor", () => {
   });
 
   it("should return status none for unknown report", async () => {
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
     const result = await sut.getReportState(mockReportId);
     assert.strictEqual(result, ExtractionStates.None);
   });
@@ -167,7 +165,7 @@ describe("BulkExtractor", () => {
       async () => ([{ imodelId: mockIModelId }])
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
 
     await sut.runReportExtractions([mockReportId]);
 
@@ -196,7 +194,7 @@ describe("BulkExtractor", () => {
       async () => ([{ imodelId: mockIModelId }, { imodelId: mockIModelId1 }])
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
 
     await sut.runReportExtractions([mockReportId]);
 
@@ -228,7 +226,7 @@ describe("BulkExtractor", () => {
       async () => ([{ imodelId: mockIModelId }, { imodelId: mockIModelId1 }])
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
 
     await sut.runReportExtractions([mockReportId]);
 
@@ -257,7 +255,7 @@ describe("BulkExtractor", () => {
       async () => ([{ imodelId: mockIModelId1 }, { imodelId: mockIModelId1 }])
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
 
     await sut.runReportExtractions([mockReportId]);
 
@@ -286,7 +284,7 @@ describe("BulkExtractor", () => {
       async () => ([{ imodelId: mockIModelId }])
     );
 
-    const sut = new BulkExtractor(mockApiConfig, mockToastCallback, mockToastCallback);
+    const sut = new BulkExtractor(new ReportsClient(), new ExtractionClient(), mockGetAccessToken, mockToastCallback, mockToastCallback);
     await sut.runReportExtractions([mockReportId]);
 
     let result = await sut.getReportState(mockReportId);