+++ packages/itwin/reports-config-widget/src/widget/components/AddMappingsModal.tsx
@@ -3,35 +3,32 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import { Modal, Table, tableFilters } from "@itwin/itwinui-react";
-import React, { useEffect, useMemo, useState } from "react";
-import type { Mapping } from "@itwin/insights-client";
-import { MappingsClient, REPORTING_BASE_PATH, ReportsClient } from "@itwin/insights-client";
+import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import type { Mapping, MappingsClient } from "@itwin/insights-client";
 import ActionPanel from "./ActionPanel";
 import "./AddMappingsModal.scss";
 import { LocalizedTablePaginator } from "./LocalizedTablePaginator";
 import type { ReportMappingAndMapping } from "./ReportMappings";
-import type { ReportsApiConfig } from "../context/ReportsApiConfigContext";
-import { useReportsApiConfig } from "../context/ReportsApiConfigContext";
+import { useReportsConfigApi } from "../context/ReportsConfigApiContext";
 import { SelectIModel } from "./SelectIModel";
 import type { CreateTypeFromInterface } from "./utils";
-import { generateUrl, handleError } from "./utils";
+import { handleError } from "./utils";
 import { ReportsConfigWidget } from "../../ReportsConfigWidget";
+import type { AccessToken } from "@itwin/core-bentley";
 
 export type MappingType = CreateTypeFromInterface<Mapping>;
 
 const fetchMappings = async (
-  setMappings: React.Dispatch<React.SetStateAction<Mapping[]>>,
+  setMappings: (mappings: Mapping[]) => void,
   iModelId: string,
-  setIsLoading: React.Dispatch<React.SetStateAction<boolean>>,
-  apiContext: ReportsApiConfig
+  setIsLoading: (isLoading: boolean) => void,
+  mappingsClient: MappingsClient,
+  getAccessToken: () => Promise<AccessToken>
 ) => {
   try {
     setIsLoading(true);
-    const mappingsClientApi = new MappingsClient(
-      generateUrl(REPORTING_BASE_PATH, apiContext.baseUrl)
-    );
-    const accessToken = await apiContext.getAccessToken();
-    const mappings = await mappingsClientApi.getMappings(
+    const accessToken = await getAccessToken();
+    const mappings = await mappingsClient.getMappings(
       accessToken,
       iModelId
     );
@@ -43,24 +40,26 @@ const fetchMappings = async (
   }
 };
 
-interface AddMappingsModalProps {
+export interface AddMappingsModalProps {
   reportId: string;
   existingMappings: ReportMappingAndMapping[];
   show: boolean;
-  returnFn: () => Promise<void>;
+  onClose: () => Promise<void>;
+  defaultIModelId?: string;
 }
 
 export const AddMappingsModal = ({
   reportId,
   existingMappings,
   show,
-  returnFn,
+  onClose,
+  defaultIModelId,
 }: AddMappingsModalProps) => {
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [selectedMappings, setSelectedMappings] = useState<Mapping[]>([]);
-  const [selectedIModelId, setSelectediModelId] = useState<string>("");
+  const selectedMappings = useRef<Mapping[]>([]);
+  const [selectedIModelId, setSelectediModelId] = useState<string | undefined>(defaultIModelId);
   const [mappings, setMappings] = useState<Mapping[]>([]);
-  const apiConfig = useReportsApiConfig();
+  const { getAccessToken, mappingsClient, reportsClient } = useReportsConfigApi();
 
   useEffect(() => {
     if (selectedIModelId) {
@@ -68,10 +67,11 @@ export const AddMappingsModal = ({
         setMappings,
         selectedIModelId,
         setIsLoading,
-        apiConfig
+        mappingsClient,
+        getAccessToken
       );
     }
-  }, [apiConfig, selectedIModelId, setIsLoading]);
+  }, [getAccessToken, mappingsClient, selectedIModelId, setIsLoading]);
 
   const mappingsColumns = useMemo(
     () => [
@@ -104,24 +104,29 @@ export const AddMappingsModal = ({
     try {
       if (!selectedIModelId) return;
       setIsLoading(true);
-      const reportsClientApi = new ReportsClient(
-        generateUrl(REPORTING_BASE_PATH, apiConfig.baseUrl)
-      );
-      const accessToken = await apiConfig.getAccessToken();
-      for (const mapping of selectedMappings) {
-        await reportsClientApi.createReportMapping(accessToken, reportId, {
+      const accessToken = await getAccessToken();
+      for (const mapping of selectedMappings.current) {
+        await reportsClient.createReportMapping(accessToken, reportId, {
           imodelId: selectedIModelId,
           mappingId: mapping.id,
         });
       }
 
-      await returnFn();
+      await onClose();
     } catch (error: any) {
       handleError(error.status);
     } finally {
       setIsLoading(false);
     }
   };
+
+  const onSelect = useCallback((selectData: Mapping[] | undefined) => {
+    if (selectData) selectedMappings.current = selectData;
+  }, []);
+
+  const tableData = useMemo(() => isLoading ? [] : mappings, [isLoading, mappings]);
+  const isRowDisabled = useCallback((rowData: MappingType) =>
+    existingMappings.some((v) => v.mappingId === rowData.id), [existingMappings]);
 
   return (
     <Modal
@@ -131,7 +136,7 @@ export const AddMappingsModal = ({
       isOpen={show}
       isDismissible={!isLoading}
       onClose={async () => {
-        await returnFn();
+        await onClose();
       }}
       style={{ display: "flex", flexDirection: "column", maxHeight: "77vh" }}
     >
@@ -141,7 +146,7 @@ export const AddMappingsModal = ({
           setSelectedIModelId={setSelectediModelId}
         />
         <Table<MappingType>
-          data={isLoading ? [] : mappings}
+          data={tableData}
           columns={mappingsColumns}
           className="rcw-add-mappings-table"
           emptyTableContent={ReportsConfigWidget.localization.getLocalizedString(
@@ -150,22 +155,17 @@ export const AddMappingsModal = ({
           isSortable
           isSelectable
           isLoading={isLoading}
-          isRowDisabled={(rowData) =>
-            existingMappings.some((v) => v.mappingId === rowData.id)
-          }
-          onSelect={(selectData: Mapping[] | undefined) => {
-            selectData && setSelectedMappings(selectData);
-          }}
+          isRowDisabled={isRowDisabled}
+          onSelect={onSelect}
           paginatorRenderer={LocalizedTablePaginator}
         />
       </div>
+      {/* Add button permanently enabled as a workaround to the warning stating that the table and parent component are being rendered at the same time. */}
       <ActionPanel
         actionLabel={ReportsConfigWidget.localization.getLocalizedString(
           "ReportsConfigWidget:Add"
         )}
         onAction={onSave}
-        onCancel={returnFn}
-        isSavingDisabled={selectedMappings.length === 0}
         isLoading={isLoading}
       />
     </Modal>