+++ packages/itwin/grouping-mapping-widget/CHANGELOG.md
@@ -1,6 +1,28 @@
 # Change Log - @itwin/grouping-mapping-widget
 
-This log was last generated on Thu, 04 May 2023 23:59:19 GMT and should not be manually modified.
+This log was last generated on Fri, 09 Jun 2023 17:54:42 GMT and should not be manually modified.
+
+## 0.17.0
+Fri, 09 Jun 2023 17:54:42 GMT
+
+### Minor changes
+
+- New version of the package targeting iTwin.js 4.x and iTwinUI 2.x!
+
+## 0.16.1
+Tue, 06 Jun 2023 10:57:06 GMT
+
+### Patches
+
+- Fixed small overflow style issue.
+
+## 0.16.0
+Mon, 15 May 2023 16:32:30 GMT
+
+### Minor changes
+
+- Refactored Copy Mapping modal UI to remove dependency on @itwin/imodel-browser-react. The widget now additionally requires the `itwins:read` scope to support the changes. 
+- @itwin/grouping-mapping-widget updated to itwinjs 3.7.4.
 
 ## 0.15.1
 Thu, 04 May 2023 23:59:19 GMT